+++ crates/oxc_ast/src/lib.rs
@@ -28,7 +28,6 @@
 #[cfg(feature = "serialize")]
 mod serialize;
 
-pub mod address;
 pub mod ast;
 mod ast_builder_impl;
 mod ast_impl;
